+++ CHANGELOG.md
@@ -7,6 +7,7 @@
   - Improve loading performances
 - Security: Support authorization and background execution for custom/plugin commands
 - New command **Reset sfdx-hardis cache** (available in help menu) to clear all cache and reload everything (after a plugin installation for example)
+- Handle case where CLI commands worker crashes
 
 ## [7.5.0] 2026-05-04