+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Security: Support authorization and background execution for custom/plugin commands
 - New command **Reset sfdx-hardis cache** (available in help menu) to clear all cache and reload everything (after a plugin installation for example)
 - Handle case where CLI commands worker crashes
+- Improve performances while listing plugins & checking versions dependencies
 
 ## [7.5.0] 2026-05-04