+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+- Pipeline View: add Zoom In / Zoom Out buttons and mouse wheel zoom support
 - Plugins architecture enhancements
   - Assign specific style to custom commands and plugins commands
   - Improve loading performances
@@ -9,6 +10,7 @@
 - New command **Reset sfdx-hardis cache** (available in help menu) to clear all cache and reload everything (after a plugin installation for example)
 - Handle case where CLI commands worker crashes
 - Improve performances while listing plugins & checking versions dependencies
+
 
 ## [7.5.0] 2026-05-04